move gageplots to gagepage new tab
