+++ src/Controllers/GagePageController.js
@@ -74,6 +74,11 @@ var StreamStats;
             }
             return RegressionType;
         }());
+        var peakValues = (function () {
+            function peakValues() {
+            }
+            return peakValues;
+        }());
         var GagePageController = (function (_super) {
             __extends(GagePageController, _super);
             function GagePageController($scope, $http, modalService, modal) {
@@ -87,6 +92,14 @@ var StreamStats;
                     displayProp: 'id'
                 };
                 _this_1.URLsToDisplay = [];
+                _this_1.floodFreq = undefined;
+                _this_1.peakDates = undefined;
+                _this_1.estPeakDates = undefined;
+                _this_1.dailyFlow = undefined;
+                _this_1.formattedFloodFreq = undefined;
+                _this_1.formattedPeakDates = undefined;
+                _this_1.formattedEstPeakDates = undefined;
+                _this_1.formattedDailyFlow = undefined;
                 $scope.vm = _this_1;
                 _this_1.modalInstance = modal;
                 _this_1.modalService = modalService;
@@ -383,9 +396,284 @@ var StreamStats;
                     }
                 }
             };
+            GagePageController.prototype.getGagePlot = function () {
+                var _this_1 = this;
+                this.gage = new GageInfo(this.modalService.modalOptions.siteid);
+                var url = configuration.baseurls.GageStatsServices + configuration.queryparams.GageStatsServicesStations + this.gage.code;
+                var request = new WiM.Services.Helpers.RequestInfo(url, true, WiM.Services.Helpers.methodType.GET, 'json');
+                this.Execute(request).then(function (response) {
+                    _this_1.gage = response.data;
+                    _this_1.getPeakInfo();
+                    _this_1.getFloodFreq();
+                    _this_1.getDailyFlow();
+                }, function (error) {
+                }).finally(function () {
+                });
+            };
+            GagePageController.prototype.getPeakInfo = function () {
+                var _this_1 = this;
+                var url = 'https://nwis.waterdata.usgs.gov/usa/nwis/peak/?format=rdb&site_no=' + this.gage.code;
+                console.log('GetPeakURL', url);
+                var request = new WiM.Services.Helpers.RequestInfo(url, true, WiM.Services.Helpers.methodType.GET, 'json');
+                this.Execute(request).then(function (response) {
+                    var peakValues = [];
+                    var estPeakValues = [];
+                    var data = response.data.split('\n').filter(function (r) { return (!r.startsWith("#") && r != ""); });
+                    var headers = data.shift().split('\t');
+                    data.shift();
+                    do {
+                        var dataRow = data.shift().split('\t');
+                        var peakObj = {
+                            agency_cd: dataRow[0],
+                            site_no: dataRow[1],
+                            peak_dt: dataRow[2],
+                            peak_va: parseInt(dataRow[4])
+                        };
+                        peakValues.push(peakObj);
+                        var estPeakObj = {
+                            agency_cd: dataRow[0],
+                            site_no: dataRow[1],
+                            peak_dt: dataRow[2].replaceAll('-00', '-01'),
+                            peak_va: parseInt(dataRow[4])
+                        };
+                        if (peakObj.peak_dt[8] + peakObj.peak_dt[9] === '00' || peakObj.peak_dt[5] + peakObj.peak_dt[6] === '00') {
+                            estPeakValues.push(estPeakObj);
+                        }
+                        ;
+                    } while (data.length > 0);
+                    var filteredArray = peakValues.filter(function (item) {
+                        return (item.peak_dt[8] + item.peak_dt[9] !== '00' || item.peak_dt[8] + item.peak_dt[9] !== '00');
+                    });
+                    _this_1.peakDates = filteredArray;
+                    _this_1.estPeakDates = estPeakValues;
+                }, function (error) {
+                }).finally(function () {
+                    _this_1.getFloodFreq();
+                });
+            };
+            GagePageController.prototype.getFloodFreq = function () {
+                var _this_1 = this;
+                var url = configuration.baseurls.GageStatsServices + configuration.queryparams.GageStatsServicesStations + this.gage.code;
+                console.log('GetFloodFreqURL', url);
+                var request = new WiM.Services.Helpers.RequestInfo(url, true, WiM.Services.Helpers.methodType.GET, 'json');
+                this.Execute(request).then(function (response) {
+                    var data = response.data;
+                    var lookup = [9, 852, 8, 4, 7, 3, 6, 1, 501, 5, 2, 500, 851, 1438, 818];
+                    var chartData = [];
+                    do {
+                        var IDs = data.statistics;
+                        for (var _i = 0, IDs_1 = IDs; _i < IDs_1.length; _i++) {
+                            var item = IDs_1[_i];
+                            if (lookup.indexOf(item.regressionTypeID) >= 0 && item.isPreferred == true) {
+                                chartData.push(item);
+                            }
+                        }
+                    } while (data.length > 0);
+                    _this_1.floodFreq = chartData;
+                }).finally(function () {
+                    _this_1.getDailyFlow();
+                });
+            };
+            GagePageController.prototype.getDailyFlow = function () {
+                var _this_1 = this;
+                var url = 'https://nwis.waterservices.usgs.gov/nwis/dv/?format=json&sites=' + this.gage.code + '&parameterCd=00060&statCd=00003&startDT=1900-01-01';
+                console.log('GetDailyFlowURL', url);
+                var request = new WiM.Services.Helpers.RequestInfo(url, true, WiM.Services.Helpers.methodType.GET, 'json');
+                this.Execute(request).then(function (response) {
+                    var data = response.data.value.timeSeries;
+                    if (data.length !== 0) {
+                        var dailyValues = data[0].values[0].value;
+                    }
+                    else {
+                        dailyValues = 0;
+                    }
+                    ;
+                    console.log('test', dailyValues);
+                    _this_1.dailyFlow = dailyValues;
+                    _this_1.formatData();
+                });
+            };
+            GagePageController.prototype.formatData = function () {
+                var _this_1 = this;
+                if (this.peakDates) {
+                    this.formattedPeakDates = [];
+                    this.peakDates.forEach(function (test) {
+                        _this_1.formattedPeakDates.push({ x: new Date(test.peak_dt), y: test.peak_va });
+                    });
+                }
+                if (this.estPeakDates) {
+                    this.formattedEstPeakDates = [];
+                    this.estPeakDates.forEach(function (test) {
+                        _this_1.formattedEstPeakDates.push({ x: new Date(test.peak_dt), y: test.peak_va });
+                    });
+                }
+                if (this.dailyFlow) {
+                    this.formattedDailyFlow = [];
+                    this.dailyFlow.forEach(function (test) {
+                        if (test.qualifiers[0] === 'A') {
+                            _this_1.formattedDailyFlow.push({ x: new Date(test.dateTime), y: parseInt(test.value) });
+                        }
+                    });
+                }
+                if (this.floodFreq) {
+                    this.formattedFloodFreq = [];
+                    var AEPColors_1 = {
+                        9: '#9A6324',
+                        852: '#800000',
+                        8: '#e6194B',
+                        818: '#ffd8b1',
+                        7: '#f58231',
+                        6: '#ffe119',
+                        5: '#bfef45',
+                        4: '#3cb44b',
+                        3: '#42d4f4',
+                        1: '#4363d8',
+                        501: '#000075',
+                        2: '#911eb4',
+                        500: '#dcbeff',
+                        851: '#fabed4',
+                        1438: '#469990'
+                    };
+                    this.floodFreq.forEach(function (floodFreqItem) {
+                        var colorIndex = floodFreqItem.regressionTypeID;
+                        var formattedName = floodFreqItem.regressionType.name.substring(0, floodFreqItem.regressionType.name.length - 18);
+                        _this_1.formattedFloodFreq.push({
+                            value: floodFreqItem.value,
+                            color: AEPColors_1[colorIndex],
+                            width: 1.5,
+                            zIndex: 4,
+                            label: { text: formattedName + '% AEP' }
+                        });
+                    });
+                    this.createAnnualFlowPlot();
+                }
+            };
+            GagePageController.prototype.createAnnualFlowPlot = function () {
+                var _this_1 = this;
+                console.log('peak value plot data', this.formattedPeakDates);
+                console.log('estimated peak plot data', this.formattedEstPeakDates);
+                console.log('daily flow plot data', this.formattedDailyFlow);
+                this.chartConfig = {
+                    chart: {
+                        height: 450,
+                        width: 800,
+                        zooming: {
+                            type: 'xy'
+                        }
+                    },
+                    title: {
+                        text: 'Annual Peak Streamflow',
+                        align: 'center'
+                    },
+                    subtitle: {
+                        text: 'Click and drag in the plot area to zoom in<br>AEP: Annual Exceedance Probability',
+                        align: 'center'
+                    },
+                    xAxis: {
+                        type: 'datetime',
+                        title: {
+                            text: 'Date'
+                        },
+                    },
+                    yAxis: {
+                        title: {
+                            text: 'Discharge (Q), in ft³/s'
+                        },
+                        plotLines: []
+                    },
+                    series: [
+                        {
+                            name: 'Daily Flow',
+                            tooltip: {
+                                headerFormat: '<b>Daily Flow</b>',
+                                pointFormatter: function () {
+                                    if (this.formattedPeakDates !== null) {
+                                        var UTCday = this.x.getUTCDate();
+                                        var year = this.x.getUTCFullYear();
+                                        var month = this.x.getUTCMonth();
+                                        month += 1;
+                                        var formattedUTCDailyDate = month + '/' + UTCday + '/' + year;
+                                        return '<br>Date: <b>' + formattedUTCDailyDate + '</b><br>Value: <b>' + this.y + ' ft³/s';
+                                    }
+                                }
+                            },
+                            turboThreshold: 0,
+                            type: 'line',
+                            color: '#add8f2',
+                            data: this.formattedDailyFlow,
+                            marker: {
+                                symbol: '',
+                                radius: 3
+                            }
+                        },
+                        {
+                            name: 'Annual Peak Streamflow',
+                            tooltip: {
+                                headerFormat: '<b>Peak Annual Flow</b>',
+                                pointFormatter: function () {
+                                    if (this.formattedPeakDates !== null) {
+                                        var waterYear = this.x.getUTCFullYear();
+                                        if (this.x.getUTCMonth() > 8) {
+                                            waterYear += 1;
+                                        }
+                                        ;
+                                        var UTCday = this.x.getUTCDate();
+                                        var year = this.x.getUTCFullYear();
+                                        var month = this.x.getUTCMonth();
+                                        month += 1;
+                                        var formattedUTCPeakDate = month + '/' + UTCday + '/' + year;
+                                        return '<br>Date: <b>' + formattedUTCPeakDate + '</b><br>Value: <b>' + this.y + ' ft³/s</b><br>Water Year: <b>' + waterYear;
+                                    }
+                                }
+                            },
+                            turboThreshold: 0,
+                            type: 'scatter',
+                            color: 'black',
+                            data: this.formattedPeakDates,
+                            marker: {
+                                symbol: 'circle',
+                                radius: 3
+                            }
+                        },
+                        {
+                            name: 'Annual Peak Streamflow (Date Estimated)',
+                            tooltip: {
+                                headerFormat: '<b>Peak Annual Flow</b>',
+                                pointFormatter: function () {
+                                    if (this.formattedPeakDates !== null) {
+                                        var waterYear = this.x.getUTCFullYear();
+                                        if (this.x.getUTCMonth() > 8) {
+                                            waterYear += 1;
+                                        }
+                                        ;
+                                        var UTCday = this.x.getUTCDate();
+                                        var year = this.x.getUTCFullYear();
+                                        var month = this.x.getUTCMonth();
+                                        month += 1;
+                                        var formattedUTCPeakDate = month + '/' + UTCday + '/' + year;
+                                        return '<br>Date (estimated): <b>' + formattedUTCPeakDate + '</b><br>Value: <b>' + this.y + ' ft³/s</b><br>Water Year: <b>' + waterYear;
+                                    }
+                                }
+                            },
+                            turboThreshold: 0,
+                            type: 'scatter',
+                            color: 'red',
+                            data: this.formattedEstPeakDates,
+                            marker: {
+                                symbol: 'square',
+                                radius: 3
+                            }
+                        }
+                    ]
+                };
+                this.formattedFloodFreq.forEach(function (formattedFloodFreqItem) {
+                    _this_1.chartConfig.yAxis.plotLines.push(formattedFloodFreqItem);
+                });
+            };
             GagePageController.prototype.init = function () {
                 this.AppVersion = configuration.version;
                 this.getGagePage();
+                this.getGagePlot();
                 this.SelectedTab = GagePageTab.GageInformation;
             };
             GagePageController.prototype.convertDateToString = function (date) {
