+++ src/Controllers/MapController.ts
@@ -455,11 +455,6 @@ module StreamStats.Controllers {
             this.modal.openModal(Services.SSModalType.e_gagepage, { 'siteid':siteid });
         }
 
-        public openGagePlot(siteid: string): void {
-            console.log('gage plot id:', siteid)
-            this.modal.openModal(Services.SSModalType.e_gageplot, { 'siteid':siteid });
-        }
-
         //Helper Methods
         //-+-+-+-+-+-+-+-+-+-+-+-
         private init(): void { 
@@ -598,8 +593,7 @@ module StreamStats.Controllers {
                                 if (item.layerName == "Streamgages" && k == "FeatureURL") {
                                     var siteNo = queryResult.properties[k].split('site_no=')[1];
                                     var SSgagepage = "vm.openGagePage('" + siteNo + "')";
-                                    var SSgageplot = "vm.openGagePlot('" + siteNo + "')";
-                                    var html = '</br><br><button ng-click="' + SSgagepage + '" type="button" class="btn-blue fullwidth">Open StreamStats Gage Page</button></br><strong>StreamStats Gage Plots: </strong><a ng-click="' + SSgageplot + '">link</a></br>';
+                                    var html = '</br><br><button ng-click="' + SSgagepage + '" type="button" class="btn-blue fullwidth">Open StreamStats Gage Page</button>';
                                     querylayers.append(html);
 
                                     //report ga event
@@ -1460,7 +1454,7 @@ module StreamStats.Controllers {
                         var gageButtonDiv = L.DomUtil.create('div', 'innerDiv');
 
                         gageButtonDiv.innerHTML = '<strong>Station ID: </strong>' + siteNo + '</br><strong>Station Name: </strong>' + feature.properties['Name'] + '</br><strong>Latitude: </strong>' + feature.geometry.coordinates[1] + '</br><strong>Longitude: </strong>' + feature.geometry.coordinates[0] + '</br><strong>Station Type</strong>: ' + feature.properties.StationType.name +
-                        '</br><br><button id="gagePageLink" type="button" class="btn-blue fullwidth">Open StreamStats Gage Page</button><strong>StreamStats Gage Plots: </strong><a id="gagePlotLink" class="' + siteNo + '">link</a><br>';
+                        '</br><br><button id="gagePageLink" type="button" class="btn-blue fullwidth">Open StreamStats Gage Page</button>';
 
                         layer.bindPopup(gageButtonDiv);
 
@@ -1487,14 +1481,6 @@ module StreamStats.Controllers {
                                 self.openGagePage(siteNo);
                             }
                         })
-
-                        L.DomEvent.on(gageButtonDiv, 'click', (event) => {
-                            const id = event.target['id'];
-                            if (id === 'gagePlotLink') {
-                                self.openGagePlot(siteNo);
-                            }
-                        })
-
 
                         layer.on('mouseover', function(e) {
                             if (self.studyArea.doSelectMapGage){
