+++ src/Services/ModalService.js
@@ -125,15 +125,6 @@ var StreamStats;
                                 backdrop: 'static',
                                 windowClass: 'windowZ',
                             };
-                        case SSModalType.e_gageplot:
-                            return {
-                                templateUrl: 'Views/gageplot.html',
-                                controller: 'StreamStats.Controllers.GagePlotController',
-                                size: 'lg',
-                                backdropClass: 'backdropZ',
-                                backdrop: 'static',
-                                windowClass: 'windowZ modal-xl',
-                            };
                         default:
                             return null;
                     }
@@ -158,7 +149,6 @@ var StreamStats;
             SSModalType[SSModalType["e_gagepage"] = 10] = "e_gagepage";
             SSModalType[SSModalType["e_nearestgages"] = 11] = "e_nearestgages";
             SSModalType[SSModalType["e_flowanywhere"] = 12] = "e_flowanywhere";
-            SSModalType[SSModalType["e_gageplot"] = 13] = "e_gageplot";
         })(SSModalType = Services.SSModalType || (Services.SSModalType = {}));
         factory.$inject = ['$modal'];
         function factory($modal) {
