--- src/Views/gageplot.html
@@ -1,43 +1,0 @@
-<!-- Modal Header-->
-<div id="print-content">
-    <div class="modal-header">
-        <button class="close hidden-print"
-                ng-click="vm.Close()">
-            &times;
-        </button>
-        <h4 class="modal-title">StreamStats Data-Collection Station Report</h4>
-    </div>
-
-<!--Modal Body-->
-        <div class="modal-body" id="printArea">
-            <div style="padding-top: 10px;">
-                <div class="panel panel-default">
-                    <div class="panel-heading">Gage Information</div>
-                        <table class="table">
-                            <tbody>
-                                <tr>
-                                    <th scope="row">USGS Station ID: </th>
-                                        <td>{{vm.gage.code}}</td>
-                                </tr>
-                                <tr>
-                                    <th scope="row">Station Name:</th>
-                                        <td>{{vm.gage.name}}</td>
-                                </tr>
-                            </tbody>
-                        </table>
-                </div>
-                    <div class="panel panel-default">
-                        <!-- <div class="panel-heading">Annual Peak Streamflow</div> -->
-                        <highchart id="chart1" config="vm.chartConfig"></highchart>
-                    </div>
-            </div>
-        </div>
-
-<!-- Footer: Close Modal Options -->
-        <div class="modal-footer hidden-print">
-            <div class="btn-group">
-                <button ng-click="vm.Close()" class="btn-black"><i class="fa fa-close"></i><span>&nbsp;&nbsp;Close</span></button>
-                <button class="btn-blue" ng-click="vm.print()"><i class="fa fa-print"></i>&nbsp;&nbsp;Print</button>
-            </div>
-        </div>
-    </div>
