+++ src/index.html
@@ -113,7 +113,6 @@
     <script src="Controllers/ExtensionModalController.js"></script>
     <script src="Controllers/FlowAnywhereController.js"></script>
     <script src="Controllers/GagePageController.js"></script>
-    <script src="Controllers/GagePlotController.js"></script>
     <script src="Controllers/NearestGagesController.js"></script>
     <!-- Directives ref-->
     <!-- endbuild -->
